mainline/common: interfaces/sensors/mainline: fix DeviceOrientationSensor to report rotation indices
WindowOrientationListener expects DEVICE_ORIENTATION scalar values
as rotation indices (0, 1, 2, 3) per the AIDL SensorType.aidl spec,
not degrees. Previously the sensor reported 0/90/180/270, causing the
framework to silently drop all non-portrait rotations due to the
range check at WindowOrientationListener.java:1148.

Rewrite ComputeOrientation() to determine which axis points toward
gravity and return the corresponding index. Also add freefall
suppression (skip when |g| < 7 m/s²) and set maxRange to 3.

+++ interfaces/sensors/mainline/DeviceOrientationSensor.h
@@ -32,11 +32,11 @@ class DeviceOrientationSensor : public ICompositeSensor {
     CompositeEvent CreateFlushCompleteEvent() const override;
 
   private:
-    float ComputeOrientation(float x, float y, float z);
+    int32_t ComputeOrientation(float x, float y, float z);
 
     CompositeSensorInfo sensor_info_;
     std::atomic_bool active_;
-    float last_orientation_;
+    int32_t last_orientation_;
     bool has_last_orientation_;
 };
